Bug 1835241 - Part 4: Make narrowing conversions explicit in GC code r=sfink
This is the most complex of these patches. I've tried to simplify existing
code and make conversions explicit. In some places I've changed types so they
make more sense (e.g. HeapThreshold::eagerAllocTrigger now returns an integer
number of bytes). I've changed some uses of float to double as it's
more standard.

I've changed some of the telemetry helper classes too. Enumeration is now
unsigned and Percentage now a double. This fits better with their uses.

Differential Revision: https://phabricator.services.mozilla.com/D181509

+++ js/src/gc/GCAPI.cpp
@@ -547,7 +547,7 @@ static bool GCBytesGetter(JSContext* cx, unsigned argc, Value* vp) {
 
 static bool MallocBytesGetter(JSContext* cx, unsigned argc, Value* vp) {
   CallArgs args = CallArgsFromVp(argc, vp);
-  double bytes = 0;
+  size_t bytes = 0;
   for (ZonesIter zone(cx->runtime(), WithAtoms); !zone.done(); zone.next()) {
     bytes += zone->mallocHeapSize.bytes();
   }
